fix: improve code style.
