+++ controllers/types.go
@@ -34,7 +34,6 @@ type TokenRequest struct {
 	SubjectTokenType    string `json:"subject_token_type"`
 	ActorToken          string `json:"actor_token"`
 	ActorTokenType      string `json:"actor_token_type"`
-	RequestedTokenType  string `json:"requested_token_type"`
 	Audience            string `json:"audience"`
 	Resource            string `json:"resource"` // RFC 8707 Resource Indicator
 }
