+++ web/src/locales/en/data.json
@@ -765,12 +765,16 @@
     "Failed to obtain MetaMask authorization": "Failed to obtain MetaMask authorization",
     "Failed to obtain Web3-Onboard authorization": "Failed to obtain Web3-Onboard authorization",
     "Forgot password?": "Forgot password?",
+    "Invalid Native SSO response": "Invalid Native SSO response",
     "LDAP": "LDAP",
     "LDAP username, Email or phone": "LDAP username, Email or phone",
     "Loading": "Loading",
     "Logging out...": "Logging out...",
     "MetaMask plugin not detected": "MetaMask plugin not detected",
     "Model loading failure": "Model loading failure",
+    "Native SSO": "Native SSO",
+    "Native SSO is ready": "Native SSO is ready",
+    "Native SSO was denied": "Native SSO was denied",
     "No account?": "No account?",
     "Or sign in with another account": "Or sign in with another account",
     "Please ensure sufficient lighting and align your face in the center of the recognition box": "Please ensure sufficient lighting and align your face in the center of the recognition box",
@@ -799,12 +803,14 @@
     "Sign in with {type}": "Sign in with {type}",
     "Signin button": "Signin button",
     "Signing in...": "Signing in...",
+    "Signed in on this device with {app}": "Signed in on this device with {app}",
     "Successfully logged in with WebAuthn credentials": "Successfully logged in with WebAuthn credentials",
     "The camera is currently in use by another webpage": "The camera is currently in use by another webpage",
     "The input is not valid Email or phone number!": "The input is not valid Email or phone number!",
     "The input is not valid Email!": "The input is not valid Email!",
     "The input is not valid phone number!": "The input is not valid phone number!",
     "To access": "To access",
+    "Use other login methods": "Use other login methods",
     "Verification code": "Verification code",
     "WeChat": "WeChat",
     "WebAuthn": "WebAuthn",
